+++ cfgov/unprocessed/js/modules/util/EventObserver.js
@@ -73,6 +73,7 @@ function EventObserver() {
   EventObserver.prototype.addEventListener = addEventListener;
   EventObserver.prototype.removeEventListener = removeEventListener;
   EventObserver.prototype.dispatchEvent = dispatchEvent;
+
   return this;
 }