+++ cfgov/unprocessed/js/molecules/GlobalSearch.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Required modules.
-var atomicCheckers = require( '../modules/util/atomic-checkers' );
+var atomicHelpers = require( '../modules/util/atomic-helpers' );
 var breakpointState = require( '../modules/util/breakpoint-state' );
 var ClearableInput = require( '../modules/ClearableInput' );
 var EventObserver = require( '../modules/util/EventObserver' );
@@ -21,8 +21,7 @@ function GlobalSearch( element ) { // eslint-disable-line max-statements, no-inl
 
   var BASE_CLASS = 'm-global-search';
 
-  var _dom =
-    atomicCheckers.validateDomElement( element, BASE_CLASS, 'GlobalSearch' );
+  var _dom = atomicHelpers.checkDom( element, BASE_CLASS, 'GlobalSearch' );
   var _triggerSel = '.' + BASE_CLASS + '_trigger';
   var _triggerDom = _dom.querySelector( _triggerSel );
   var _flyoutMenu = new FlyoutMenu( _dom ).init();